requirements.txt
ccxt==2.8.75
python-dotenv==1.0.0
numpy==1.26.5
pandas==2.2.2
ta==0.11.0
scikit-learn==1.4.2
websockets==11.0.3
